refactor(event): Refactor events component
