+++ src/events/event-record.ts
@@ -5,9 +5,9 @@
  * Licensed under MIT License (see https://github.com/regal/regal)
  */
 
-import { PropertyChange, PropertyOperation } from "../agents";
+import { PropertyChange } from "../agents";
 import { OutputLine } from "../output";
-import { noop, TrackedEvent } from "./event-model";
+import { TrackedEvent } from "./event-types";
 
 /** Event ID for untracked `EventFunction`s. */
 export const DEFAULT_EVENT_ID: number = 0;
@@ -17,44 +17,33 @@ export const DEFAULT_EVENT_NAME: string = "DEFAULT";
 /**
  * Record of a `TrackedEvent`'s effects in a game cycle.
  */
-export class EventRecord {
+export interface EventRecord {
+    /* The event's unique numeric ID. */
+    id: number;
+
+    /* The event's name. */
+    name: string;
+
+    /* The event's `TrackedEvent`. */
+    func: TrackedEvent;
+
     /** The IDs of the `OutputLine`s emitted by the event. */
-    public output?: number[];
+    output?: number[];
+
     /** The ID of the event that caused this event. */
-    public causedBy?: number;
+    causedBy?: number;
+
     /** The IDs of the events that were caused by this event. */
-    public caused?: number[];
+    caused?: number[];
+
     /** The records of all changes to registered agents that were caused by this event. */
-    public changes?: PropertyChange[];
-
-    /** Default `EventRecord` for untracked `EventFunction`s. */
-    public static get default() {
-        return new EventRecord();
-    }
-
-    /**
-     * Constructs a new `EventRecord`.
-     *
-     * @param id The event's unique numeric ID (optional).
-     * @param name The event's name (optional).
-     * @param func The event's `TrackedEvent`. Defaults to `noop`.
-     */
-    constructor(
-        public id: number = DEFAULT_EVENT_ID,
-        public name: string = DEFAULT_EVENT_NAME,
-        public func: TrackedEvent = noop
-    ) {}
+    changes?: PropertyChange[];
 
     /**
      * Appends a reference to an `OutputLine` to the `EventRecord`'s output log.
      * @param line The line of output emitted by the event.
      */
-    public trackOutputWrite(line: OutputLine): void {
-        if (this.output === undefined) {
-            this.output = [];
-        }
-        this.output.push(line.id);
-    }
+    trackOutputWrite(line: OutputLine): void;
 
     /**
      * Creates an association between this `EventRecord` and the caused events,
@@ -63,23 +52,11 @@ export class EventRecord {
      *
      * @param events The events that were caused by this event.
      */
-    public trackCausedEvent(...events: EventRecord[]): void {
-        if (this.caused === undefined) {
-            this.caused = [];
-        }
-        this.caused.push(...events.map(e => e.id));
-        events.forEach(e => (e.causedBy = this.id));
-    }
+    trackCausedEvent(...events: EventRecord[]): void;
 
     /**
      * Adds a record of a single change to a registered agent's property to the
      * `EventRecord`'s `changes` property.
      */
-    public trackChange(propChange: PropertyChange): void {
-        if (this.changes === undefined) {
-            this.changes = [];
-        }
-
-        this.changes.push(propChange);
-    }
+    trackChange(propChange: PropertyChange): void;
 }
